"Implement Home screen UI components and ViewModel for user rooms and link creation"
In this commit, the following classes were added to the presentation layer:

HomeUIEvent: Represents the UI events specific to the Home screen.
HomeViewModel: Manages the logic and data for the Home screen, including fetching user rooms and creating a link room.
HomeScreen: Implements the UI for the Home screen, including user interface elements and interactions.
HomeUiState: Represents the different UI states for the Home screen.
InputFieldUiState: Represents the different UI states for the Input Fields.
Constants: including presentation layer constants
and some utilities such as
DateAndTimeUtils: convert time in milliseconds to timestamp
RoomItemPositionUtil: set position of item in list